Fix for force-unwrap crash in DFURemoteError.with(code:)
We found this crash in nRF Connect. Filip believes this was a user testing their new firmware, because all error codes are supposed to be present. In any case, we decided this force-unwrap could be avoided, so if an error is not found, we presume the DFU device's response is not correct.